% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sunburst.R
\name{add_remain_and_diff_cols}
\alias{add_remain_and_diff_cols}
\title{Add Remain and Diff Columns to a Data Frame}
\usage{
add_remain_and_diff_cols(btns_df, totalPathways)
}
\arguments{
\item{btns_df}{A data frame with a \code{Remain} column of numeric counts.}

\item{totalPathways}{A numeric value for the total number of pathways.}
}
\value{
A data frame with formatted \code{Remain} and \code{Diff} columns.
}
\description{
Add Remain and Diff Columns to a Data Frame
}
