% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sunburst.R
\name{customActionButton}
\alias{customActionButton}
\title{Create a Colored Action Button}
\usage{
customActionButton(inputId, label, color, font_size = "10px")
}
\arguments{
\item{inputId}{The input slot for the button.}

\item{label}{The button label.}

\item{color}{A CSS color string for the button background.}

\item{font_size}{A CSS font size string.}
}
\value{
A \code{shiny::actionButton} with custom styling.
}
\description{
Create a Colored Action Button
}
