% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_methods.R
\name{getPathwayGroupDatatable}
\alias{getPathwayGroupDatatable}
\title{Get Pathway Group Data Table}
\usage{
getPathwayGroupDatatable(id, pathwayAnalysisDTO, pathLength)
}
\arguments{
\item{id}{The namespaced widget output ID.}

\item{pathwayAnalysisDTO}{A list containing the pathway analysis data.}

\item{pathLength}{An integer specifying the pathway length.}
}
\value{
The widget element ID (invisibly).
}
\description{
Sends a custom message to the sunburst widget to retrieve pathway group data.
}
