% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sunburst.R
\name{match_color}
\alias{match_color}
\title{Match Event Names to Colors}
\usage{
match_color(x, eventCodes)
}
\arguments{
\item{x}{A character string of comma-separated event names.}

\item{eventCodes}{A data frame with \code{name} and \code{color} columns.}
}
\value{
A character string of HTML action buttons.
}
\description{
Match Event Names to Colors
}
