% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburstShinyWidget.R
\name{sunburstShinyWidget}
\alias{sunburstShinyWidget}
\title{Create a Sunburst Shiny Widget}
\usage{
sunburstShinyWidget(
  data,
  design,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A list containing the chart data.}

\item{design}{A list containing the design configuration.}

\item{width}{Widget width (a valid CSS unit or a number).}

\item{height}{Widget height (a valid CSS unit or a number).}

\item{elementId}{An optional element ID for the widget.}
}
\value{
An \code{htmlwidget} object.
}
\description{
Creates an interactive sunburst plot as an HTML widget.
}
